Move most of the configuration files into .config/
XDG Base Directory Specification says that the default directory
for storing configuration should be ~/.config/
(http://standards.freedesktop.org/basedir-spec/latest/index.html).

Unfortunately many programs don't respect this standard (for example
bash itself) - if that wasn't the case, I would probably transform
this repository so that it would simply track the contents of .config/
instead of the whole HOME.  However, I'm trying to move as much into
.config/ as possible.

Because of that I have changed the path from which the 'dotfiles'
command expects to read the localization of the repository database
from ~/dotfiles-repo-path to ~/.config/dotfiles-git-dir - you have
to move your file accordingly.