Git logs: don't use "git lu" as the base for other logs
The problem with "git lu" is that it will include currently checked-out
branch and it's upstream even if called with another branch argument.

It would be best to have a command that logs only a specified branch
with that branch's upstream, AND use current branch as the default,
but I didn't manage to do that yet.