Symlink dotfiles instead of using "detached git dir"
Previously, dotfiles were configured so that the git working directory
was actually $HOME, but git dir was separated from it (for example, it
could be in ~/.dotfiles) so that one couldn't easily mess up his $HOME
by running a git command inside it without meaning to.  To manage the
repo, one had to use special `dotfiles` alias.

However, I have decided that this setup is too complicated - at least
for the sensible-dotfiles repository.  After all, the purpose of this
repo is to be a *simple* collection of most common CLI adjustments.

From now on, the installation script will simply symlink all files from
the repository to $HOME, exactly mirroring their locations.  In fact,
it seems that this is how most people manage their dotfile repositories.

Pros:
- overall simpler, needs less explanations
- no need for special `dotfiles` alias
- no need for complicated .gitignore
- no need for a config file storing repo path
- connection between files in $HOME and repo is explicit
- README and installation script won't appear in $HOME

Cons:
- one has to run the symlinking script again whenever new new files are
  added to the repo
- switching branches (admittedly rare) may leave you with broken
  symlinks
- some 3rd-party program may mess up the symlinks (?)