Don't use less --RAW-CONTROL-CHARS by default (fix #3)
I didn't realize that control characters are powerful enough that
parsing them can be a security vulnerability (although the likeliness of
such attack is low, and it's not only less that is affected so that it's
not really feasible to fully protect yourself from this issue):
http://openwall.com/lists/oss-security/2015/09/17/5
http://turbochaos.blogspot.ca/2014/08/journalctl-terminal-escape-injection.html

Because of this I remove --RAW-CONTROL-CHARS from less' "default settings".
Thanks @msoczewka for the info!