OSX compatibility - use .bash_profile
Bash has weird preferences with regard to config files.  When it's a
non-login shell, it reads user's .bashrc - easy peasy.  But for login
shells (which is the case on Macs most of the time), it will read the
first of .bash_profile, .bash_login, .profile.  In particular, it will
not read .profile when .bash_profile exists, so to make sure that
sensible-dotfiles' configuration is loaded on Macs, I have to put it
in .bash_profile rather than .profile.

In this case, I can remove [ -n "$BASH_VERSION" ] check, as no other
shell than bash will read .bash_profile.